# apps/api/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NEXAMed API",
    "version": "1.0.0",
    "description": "Sinh tự động từ packages/shared — không sửa tay."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/api/v1/patients": {
      "post": {
        "tags": [
          "patient"
        ],
        "summary": "Tạo hồ sơ bệnh nhân (PAT-01)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "fullName": {
                    "type": "string",
                    "minLength": 1
                  },
                  "dob": {
                    "type": "string",
                    "format": "date"
                  },
                  "gender": {
                    "type": "string",
                    "enum": [
                      "male",
                      "female",
                      "other"
                    ]
                  },
                  "phone": {
                    "type": "string",
                    "minLength": 8,
                    "maxLength": 15
                  },
                  "nationalId": {
                    "type": "string",
                    "minLength": 9,
                    "maxLength": 12
                  },
                  "nationalIdIssuedAt": {
                    "type": "string",
                    "format": "date"
                  },
                  "nationalIdIssuedPlace": {
                    "type": "string",
                    "minLength": 1
                  },
                  "ethnicity": {
                    "type": "string",
                    "minLength": 1
                  },
                  "nationality": {
                    "type": "string",
                    "minLength": 1
                  },
                  "occupation": {
                    "type": "string",
                    "minLength": 1
                  },
                  "insuranceNumber": {
                    "type": "string",
                    "minLength": 1
                  },
                  "address": {
                    "type": "object",
                    "properties": {
                      "street": {
                        "type": "string",
                        "minLength": 1
                      },
                      "ward": {
                        "type": "string",
                        "minLength": 1
                      },
                      "neighborhood": {
                        "type": "string",
                        "minLength": 1
                      },
                      "district": {
                        "type": "string",
                        "minLength": 1
                      },
                      "province": {
                        "type": "string",
                        "minLength": 1
                      }
                    }
                  },
                  "allergyNote": {
                    "type": "string"
                  },
                  "relativeFullName": {
                    "type": "string",
                    "minLength": 1
                  },
                  "relativeRelationship": {
                    "type": "string",
                    "minLength": 1
                  },
                  "relativePhone": {
                    "type": "string",
                    "minLength": 1
                  },
                  "relativeAddress": {
                    "type": "string",
                    "minLength": 1
                  }
                },
                "required": [
                  "fullName",
                  "dob",
                  "gender",
                  "phone"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tạo thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "patientCode": {
                          "type": "string"
                        },
                        "fullName": {
                          "type": "string"
                        },
                        "dob": {
                          "type": "string"
                        },
                        "gender": {
                          "type": "string",
                          "enum": [
                            "male",
                            "female",
                            "other"
                          ]
                        },
                        "phone": {
                          "type": "string"
                        },
                        "hasNationalId": {
                          "type": "boolean"
                        },
                        "nationalIdMasked": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "address": {
                          "type": [
                            "object",
                            "null"
                          ],
                          "properties": {
                            "street": {
                              "type": "string",
                              "minLength": 1
                            },
                            "ward": {
                              "type": "string",
                              "minLength": 1
                            },
                            "neighborhood": {
                              "type": "string",
                              "minLength": 1
                            },
                            "district": {
                              "type": "string",
                              "minLength": 1
                            },
                            "province": {
                              "type": "string",
                              "minLength": 1
                            }
                          }
                        },
                        "version": {
                          "type": "integer"
                        },
                        "nationalId": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "nationalIdIssuedAt": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "nationalIdIssuedPlace": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "ethnicity": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "nationality": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "occupation": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "insuranceNumber": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "allergyNote": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "relativeFullName": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "relativeRelationship": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "relativePhone": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "relativeAddress": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "photoUrl": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "mergedIntoId": {
                          "type": [
                            "string",
                            "null"
                          ],
                          "format": "uuid"
                        }
                      },
                      "required": [
                        "id",
                        "patientCode",
                        "fullName",
                        "dob",
                        "gender",
                        "phone",
                        "hasNationalId",
                        "nationalIdMasked",
                        "address",
                        "version",
                        "nationalId",
                        "nationalIdIssuedAt",
                        "nationalIdIssuedPlace",
                        "ethnicity",
                        "nationality",
                        "occupation",
                        "insuranceNumber",
                        "allergyNote",
                        "relativeFullName",
                        "relativeRelationship",
                        "relativePhone",
                        "relativeAddress",
                        "photoUrl",
                        "mergedIntoId"
                      ]
                    },
                    "meta": {
                      "type": "object",
                      "properties": {}
                    }
                  },
                  "required": [
                    "data",
                    "meta"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Thiếu hoặc sai access token",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {}
                      },
                      "required": [
                        "code",
                        "message"
                      ]
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "403": {
            "description": "Không có quyền patient.create",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {}
                      },
                      "required": [
                        "code",
                        "message"
                      ]
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "409": {
            "description": "Trùng số CCCD/CMND với bệnh nhân khác trong tenant",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {}
                      },
                      "required": [
                        "code",
                        "message"
                      ]
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "patient"
        ],
        "summary": "Danh sách bệnh nhân, phân trang cursor (PAT-02)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": false,
            "name": "cursor",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            },
            "required": false,
            "name": "limit",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 100
            },
            "required": false,
            "name": "q",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "items": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string",
                                "format": "uuid"
                              },
                              "patientCode": {
                                "type": "string"
                              },
                              "fullName": {
                                "type": "string"
                              },
                              "dob": {
                                "type": "string"
                              },
                              "gender": {
                                "type": "string",
                                "enum": [
                                  "male",
                                  "female",
                                  "other"
                                ]
                              },
                              "phone": {
                                "type": "string"
                              },
                              "hasNationalId": {
                                "type": "boolean"
                              },
                              "nationalIdMasked": {
                                "type": [
                                  "string",
                                  "null"
                                ]
                              },
                              "address": {
                                "type": [
                                  "object",
                                  "null"
                                ],
                                "properties": {
                                  "street": {
                                    "type": "string",
                                    "minLength": 1
                                  },
                                  "ward": {
                                    "type": "string",
                                    "minLength": 1
                                  },
                                  "neighborhood": {
                                    "type": "string",
                                    "minLength": 1
                                  },
                                  "district": {
                                    "type": "string",
                                    "minLength": 1
                                  },
                                  "province": {
                                    "type": "string",
                                    "minLength": 1
                                  }
                                }
                              },
                              "version": {
                                "type": "integer"
                              }
                            },
                            "required": [
                              "id",
                              "patientCode",
                              "fullName",
                              "dob",
                              "gender",
                              "phone",
                              "hasNationalId",
                              "nationalIdMasked",
                              "address",
                              "version"
                            ]
                          }
                        },
                        "nextCursor": {
                          "type": [
                            "string",
                            "null"
                          ]
                        }
                      },
                      "required": [
                        "items",
                        "nextCursor"
                      ]
                    },
                    "meta": {
                      "type": "object",
                      "properties": {}
                    }
                  },
                  "required": [
                    "data",
                    "meta"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Thiếu hoặc sai access token",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {}
                      },
                      "required": [
                        "code",
                        "message"
                      ]
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "403": {
            "description": "Không có quyền patient.read",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {}
                      },
                      "required": [
                        "code",
                        "message"
                      ]
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}